chore(arch): move go code into pkg
