+++ .github/workflows/golangci.yml
@@ -35,9 +35,6 @@ jobs:
         path: ./target
         key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('./Cargo.lock') }}
         
-    # Build packages
-    - name: Build
-      run: make
     # Lint code
     - name: Lint
       uses: golangci/golangci-lint-action@v1
