+++ pkg/internal/util/collection.go
