+++ pkg/internal/util/encode.go
